+++ src/app/app.component.ts
@@ -1,12 +1,12 @@
 import { ChangeDetectorRef, Component, OnDestroy, ViewChild } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 import { MediaMatcher } from '@angular/cdk/layout';
+import { DomSanitizer } from '@angular/platform-browser';
 import { filter } from 'rxjs/operators';
 
 import { MatSidenav } from '@angular/material/sidenav';
 import { MatIconRegistry } from '@angular/material/icon';
-import { DomSanitizer } from '@angular/platform-browser';
-import { GlobalVariableService } from './services/global-variable.service';
+import { GlobalVariableService, LinkInfo } from './services/global-variable.service';
 
 @Component({
   selector: 'app-root',
@@ -17,9 +17,14 @@ export class AppComponent implements OnDestroy {
   title = 'angular-sample';
 
   @ViewChild('leftSidenav') leftSidenav!: MatSidenav;
-  isSidenavVisible: boolean = false;
-  mobileQuery: MediaQueryList;
   private _mobileQueryListener: (e: MediaQueryListEvent) => void;
+
+  get isMobile(): boolean {
+    return this.gvs.isMobile;
+  }
+  get linkInfos(): LinkInfo[] {
+    return this.gvs.LinkInfos;
+  }
 
   constructor(
     router: Router,
@@ -27,19 +32,16 @@ export class AppComponent implements OnDestroy {
     media: MediaMatcher,
     matIconRegistry: MatIconRegistry,
     domSanitizer: DomSanitizer,
-    gvs: GlobalVariableService) {
+    private gvs: GlobalVariableService) {
 
     router.events.pipe(
       filter(e => e instanceof NavigationEnd)
     ).subscribe(event => this.updateSidenavVisibilty(event as NavigationEnd));
 
-    this.mobileQuery = media.matchMedia('(max-width: 600px)');
-    this._mobileQueryListener = (e: MediaQueryListEvent) => {
-      gvs.isMobile = e.matches;
-      changeDetectorRef.detectChanges();
-    };
-    gvs.isMobile = this.mobileQuery.matches;
-    this.mobileQuery.addEventListener('change', this._mobileQueryListener);
+    // 20210909: detect window width
+    gvs.mobileQuery = media.matchMedia('(max-width: 600px)');
+    this._mobileQueryListener = (e: MediaQueryListEvent) => changeDetectorRef.detectChanges();
+    gvs.mobileQuery.addEventListener('change', this._mobileQueryListener);
 
     matIconRegistry.addSvgIcon(
       "githubicon",
@@ -48,7 +50,7 @@ export class AppComponent implements OnDestroy {
   }
 
   ngOnDestroy(): void {
-    this.mobileQuery.removeEventListener('change', this._mobileQueryListener);
+    this.gvs.mobileQuery.removeEventListener('change', this._mobileQueryListener);
   }
 
   ngAfterViewInit(): void {
@@ -65,10 +67,7 @@ export class AppComponent implements OnDestroy {
   updateSidenavVisibilty(event: NavigationEnd) {
     // console.log(event);
 
-    let isSidenavVisible = this.getSidenavVisibilty(event.urlAfterRedirects, '/investment');
-
-    this.isSidenavVisible = isSidenavVisible;
-    this.leftSidenav.opened = isSidenavVisible;
+    this.leftSidenav.opened = !this.isMobile && this.getSidenavVisibilty(event.urlAfterRedirects, '/sidenav');
   }
 
   getSidenavVisibilty(url: string, prefix: string) {