+++ src/app/pages/sidenav-link-item/sidenav-link-item.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SidenavComponent } from './sidenav.component';
+import { SidenavLinkItemComponent } from './sidenav-link-item.component';
 
-describe('SidenavComponent', () => {
-  let component: SidenavComponent;
-  let fixture: ComponentFixture<SidenavComponent>;
+describe('SidenavLinkItemComponent', () => {
+  let component: SidenavLinkItemComponent;
+  let fixture: ComponentFixture<SidenavLinkItemComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ SidenavComponent ]
+      declarations: [ SidenavLinkItemComponent ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(SidenavComponent);
+    fixture = TestBed.createComponent(SidenavLinkItemComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });