+++ src/app/pages/sidenav-link-item/sidenav-link-item.component.ts
@@ -1,0 +1,23 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { GlobalVariableService, LinkInfo } from 'src/app/services/global-variable.service';
+
+@Component({
+  selector: 'app-sidenav-link-item',
+  templateUrl: './sidenav-link-item.component.html',
+  styleUrls: ['./sidenav-link-item.component.scss']
+})
+export class SidenavLinkItemComponent implements OnInit {
+
+  @Input() linkInfos!: LinkInfo[];
+  @Input() hiddenLevels!: number[];
+  @Input() currentLevel!: number;
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+  isHiddenLevel() {
+    return this.hiddenLevels.includes(this.currentLevel);
+  }
+}