+++ src/app/pages/toolbar/toolbar.component.ts
@@ -1,5 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Router, UrlTree } from '@angular/router';
 import { trigger, style, animate, transition, state } from '@angular/animations';
+import { GlobalVariableService, LinkInfo } from 'src/app/services/global-variable.service';
 
 @Component({
   selector: 'app-toolbar',
@@ -27,10 +29,40 @@ import { trigger, style, animate, transition, state } from '@angular/animations'
 })
 export class ToolbarComponent implements OnInit {
 
-  @Input() isMenuBtnVisible: boolean = false;
   @Output() menuClickEvent = new EventEmitter();
 
-  constructor() { }
+  get isMobile(): boolean {
+    return this.gvs.isMobile;
+  }
+  get linkInfos(): LinkInfo[] {
+    return this.gvs.LinkInfos;
+  }
+
+  get isShownMenu(): boolean {
+    // todo: 20210909: figure out IsActiveMatchOptions params
+
+    let isShownMenu: boolean = false;
+    for (let linkInfo of this.linkInfos) {
+
+      let url: string | UrlTree = this.router.createUrlTree(linkInfo.url);
+      let isActive = this.router.isActive(url, {
+        matrixParams: 'subset',
+        queryParams: 'subset',
+        paths: 'subset',
+        fragment: 'exact'
+      })
+
+      if (isActive && linkInfo.subLink !== undefined) {
+        isShownMenu = true;
+      }
+    }
+    return this.isMobile || isShownMenu;
+  }
+  get displayedLinkInfos(): LinkInfo[] {
+    return this.isMobile ? [] : this.linkInfos;
+  }
+
+  constructor(private router: Router, private gvs: GlobalVariableService) { }
 
   ngOnInit(): void {
   }